koala/koala_server/requirements.txt
Flask==1.0
psycopg2==2.8.1
sqlalchemy==1.3.2
sqlalchemy_utils==0.33.11
flask-sqlalchemy==2.3.2
flask-restful==0.3.7
flask-cors==3.0.7
requests==2.21.0
pandas==0.24.2
xlsxwriter==1.1.6
flask_caching==1.7.1
validators==0.12.4
Pillow==6.0.0
numpy==1.16.2
openpyxl==2.6.2
flask_httpauth==3.2.4
itsdangerous==1.1.0
flask-migrate==2.4.0
flask-script==2.0.6
xlrd==1.2.0

